Add files via upload
With millions of people possessing cars, monitoring cars is an important matter. One of the issues is the gas pollution, with cars exhausting a lot of gases, in addition to that there is no monitoring or collecting of information about the quantity of gases being exhausted. In this project, I present a way to monitor a car location using GPS module and gas sensor to collect gas pollution data.